# public/swagger/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "version": "v1",
      "title": "Hefesto Symfony",
      "description": "Hefesto Symfony",
      "contact": {
        "name": "Henrique Figueiredo de Souza"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "servers": [
      {
        "url": "http://localhost:8080/api/v1",
        "description": "Local server"
      }
    ],
...
    "paths": {
...
      "/admParameterCategory": {
        "get": {
          "tags": [
            "AdmParameterCategory"
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdmParameterCategory"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdmParameterCategory"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdmParameterCategory"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "AdmParameterCategory"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmParameterCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmParameterCategory"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmParameterCategory"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/AdmParameterCategory"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AdmParameterCategory"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AdmParameterCategory"
                  }
                }
              }
            }
          }
        }
      },
      "/admParameterCategory/{id}": {
        "get": {
          "tags": [
            "AdmParameterCategory"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/AdmParameterCategory"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AdmParameterCategory"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AdmParameterCategory"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "AdmParameterCategory"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmParameterCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmParameterCategory"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmParameterCategory"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "AdmParameterCategory"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "AdmParameterCategory": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "order": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }